输入标题时，实现模糊搜索，显示可能存在的标题名字

+++ SearchEbook/Model/BookDetialInfo.cs
@@ -32,7 +32,7 @@ namespace SearchEbook.Model
         public int followerCount { get; set; }
         public int wordCount { get; set; }
         public int serializeWordCount { get; set; }
-        public int retentionRatio { get; set; }
+        public double retentionRatio { get; set; }
         public DateTime updated { get; set; }
         public bool isSerial { get; set; }
         public int chaptersCount { get; set; }
